TEI exporter WIP v0.5:
c(4) = comparing multiple hypernodes of unequal length

note: c(4) is operational on hypernode of the same size; more precisely, it works on graphs where ommited element are explicitly described on the graph
note: a section node (main graph) can have multiple hypernodes. To indicate the main/lemma hypernode, which is the hypernode that is grouping node from the lemma section,
the user have to add the mention "main" in the note field when a hypernode is created.

todo : add a check box "main" in the hypernode dialog box to allow a simple hn polarisation/lemmatization
todo : ideally add a "ranking"/"order" property to the HAS_HYPERNODE relation to garanty sorted nodes when nodes are fetched. The property should be set, API side,
when a hypernode is created.
todo : add a "omission" relation for describing relations between node and empty nodes
todo : add a "remove" empty node to restore graph to the state before a node is added with the "append"(a) function
todo : add a "attach" function to reverse "detach" (d) operations

- c(4) prototype implemented
- edit top hypernode selection mode
- fix hypernode node order